Added placeholder files for autotools
``Building'' Liza will not be required to use it; see ease.js for the style
of building that will be used here. In particular: the build process will be
used for minification and the creation of distribution archives, but one can
simply include the source files (using, say, node.js) to use it without
building.