[DEV-7109] Fix: Interpret __retry and set __rate_pending accurately
__retrys were found to contain nested arrays which would cause false
positives when assuming they were only one level deep. Now the data are
being extracted properly by flattening the data structure. Based on
these data, __rate_pending was being set to the count of retries.
However, this value needed to be wrapped in an array to conform with
our data standards.

Co-Authored-By: Austin Schaffer <austin.schaffer@ryansg.com>