[DEV-7395] TableGroupUi calls GroupContext on show and hide with TableCellFieldContext
Detaching Table cells proved to be an issue because the
delete row cell was moved to the first cell instead of the last
because it is not an exclusive field, so the getNextElement
function did not work as expected. Now the delete row cell
is put back as the last child of the row. Also, TableGroupUi
overrides doShowField and doHideField so it was never calling
the GroupContext. We want a FieldContext for every field which
would only happen if GroupContext is called for those fields.