[DEV-7533] Remove focus on first input element
Focusing on the first input element does not work properly
already since it doesn't take into account hidden fields
like the test indication checkbox that is hidden for
external users.

When fields are detached from the DOM and then re-attached
with Cmatch, the first non-cmatch field was focused on step load,
which can cause unintended scrolling to a field further
down the step.