[DEV-7591] Correctly select the last eligible tab
The previous implementation of TabbedBlockGroupUi._getLastEligibleTab()
counted the number of eligible tabs without accounting for how many tabs
were not eligible. This lead to situations in which the tab selected was
not the last eligible tab.

In addition, the TabbedBlockGroupUI was updated to always use the jQuery
object provided at instantiation. This helps with unit testing and
avoids potential discrepancies.