[DEV-7535] Refactor: Create state manager for groups
In the interest of the single responsibility principle, a group need not
manage its own state.

GridCellGroup requires tracking of a pending state, similar to how
TabbedBlockGroup uses blockFlags. With both use cases being so similar,
it made sense to extract the related code into a new class that manages
this state for the groups.