[DEV-8180] Remove session redirect support
The session redirect was causing issues with
the Start Quote page on the Connector that would
take a user to an old quote instead of a new
quote. This had to potential to merge data from
a new quote into an old quote with the Acord to
Quote feature.

This redirect logic should be rewritten to
include the quote number in the URL instead
of store the quote number in the session data.