[DEV-8167] Introduce concept of a dirty quote
This will be used in addition to the dirty bucket indicator
to identify quotes which require step saves. Navigation is
alreadying using quote.isDirty() to control access to steps.
Now that we have autosave, we need to force steps to save
even if the bucket is not commited and bucket is not dirty, this
is necessary enforce step level assertions.