[DEV-7268] GeneralStepUi: Do not update answer fields if step not active
This fixes a race condition where you can switch from a step
that is dirty to another step, forcing a stepSave which then
tries to update displays on a step that is no longer active.

The GeneralStepUi setActive method was being called after
answerDataUpdate, which caused the race condition to appear.