[DEV-8610] [DEV-8399] Upgrade eslint to 4.1
There are a number of fixes that had to be made or ignored from new rules,
depending on the effort involved.

Certain files failed to parse, for various reasons; they have been added for
now to .eslintignore.  I will check back when I have more time.  We will not
be changing our TS to accomodate a failure by eslint; if needed, we can fix
the problem and submit a patch upstream.