[DEV-8378] Create 'Dullahan' daemon, and indication endpoint
Dullahan is an Irish spirit who carries his own head. The name is in
reference to the detached nature of this process from a UI. The
Dullahan daemon starts an Express application that handles different
endpoints associated with a detached rating system.

These changes also include ancillary wrappers around the fetch API and
the Express router, as well as a controller for indication routes, and
the daemon itself.