[DEV-7738] SideTableGroupUi does not detach table and removes multiple index support
SideTableGroupUi will now function with DOM Performance changes,
but with very limited support. The group will no longer support
multiple indexes.

GroupUi requires a unique fieldContentParent for each index,
yet the fields in this group can have different parent row elements
for the same index. This means that any cmatch fields that need
to show and hide on this group will not work properly.

TableGroupUi and TabbedBlockGroupUi should be heavily favored over
SideTableGroupUi, which should be removed as soon as it is no
longer being used.