[DEV-8551] Fix race condition with UI preRender event
The ElementStyler needs the current step content to query
for elements that are added to the DOM. In some circumstances,
when a user navigates back to a step and a new display has been added,
the ElementStyler context was the previous step which caused it
not to find the new display, causing an error that was displayed
to the user.

This appears to fix a UI race condition bug - the event emitted for
EVENT_PRE_RENDER_STEP was happening after groups had started the
preRender process, which when navigating back and forth between steps
often required a handleIndexChange, which will run revokeStyle on a
DomField, but the ElementStyler was not updated yet with the correct
$context (the step div). Moving the event before preRender method
fixed it so ElementStyler had the correct $context sooner.