[DEV-8616] MongoServerDao#pullQuote: Promise and rejection on error
This no longer uses a callback.  Further, it will actually fail if there is
an error pulling back the document.  A tad bit frightening that it never
did, though it doesn't seem to have caused any practical issues for the past
decade.

This doesn't include a corresponding test because the Mongo driver will be
updated soon and the API has changed; I'll worry about it then.