[DEV-8616] Initialize new quotes before triggering redirect
How this worked previously was nonsense, and I'm surprised it lasted this
long.  Creating a new quote was a consequence of how the system worked
previously---we took advantage of the fact that requesting quote 0 would
cause a redirect to the next available quote id, and then, when that quote
was requested, we would create a new one.  This resulted in unnecessarily
complicated logic.

Instead, when quote 0 is requested (which is invalid), we create a new quote
_before_ redirecting the user to it; that allows us to simply fail if a
quote cannot be loaded when requested.

This is still a mess, and errors creating a quote are still not being
handled (!); this is an incremental step in the right direction given the
little time that I have available.

This should also fix DEV-8406, which under certain circumstances permits
quote 0 to be saved to the database, which makes it impossible to create new
quotes using that scheme.

Another new regarding Mocha and `this.timeout(0)`---that caused a debugging
headache when it caused Mocha to silently terminate with a 0 exit code
instead of causing a failure of some sort.  It also ran no tests after it.