[DEV-8616] fp-ts: New dependency
This will be used by future commits.

I've wanted to introduce a functional library into Liza for quite some time,
and TS has enabled the creation of a fairly robust implementation.  Others
on the team did the research on this (Joe and Anthony, primarily).

While it does have some shortcoming and awkward limitations due to TS, and
it is heavy on terminology, it does seem like a good library so far.