[DEV-8616] Sever#_loadDocumentIntoQuote: Extract session portion
This begins a transition into a functional module, which I'll commit
incrementally to show how it evolves.  It is tightly coupled at the moment,
and probably will remain that way until it's extracted out of Server
entirely.  This is partially due to my availability.

This is the first use of fp-ts here.