[DEV-8616] Server#_getDefaultBucket: Extract
This extracts the logic into `loader`.  This is once again a transitionary
commit; see previous commits for more information.

This change is notable in that the `#initQuote` method is now fully glue
code for functions in the loader, and so itself can now be extracted in its
entirety.

Also noteworthy is that I intend to merge ProgramInit itself into `loader`,
since it does not make sense to have that logic stand on its own.