[DEV-8616] ProgramQuoteCleaner#clean: Return Promise
This will aid in a transition to the loader.

Further, it appears that it never actually returns any errors...and there's
a bunch of complex code around whether cleaning fails.  This can be cleaned
up a ton, which would be great, since version migration is some of the most
confusing (and therefore dangerous) code.