[DEV-8616] Server#_checkQuotePver: Simplify quote cleaning integration
This gets rid of the nasty hook-based system and instead injects the
dependency.  While this does add to the dependencies of Server, which is
bad, the overall effect is much cleaner, more understandable code.  The
dependency will be removed in the future as things are moved into the
loader.

I decided to leave `clean` async since, as it becomes more intensive, we may
wish to take advantage of that, at least as long as it's on the same
thread.  Further, other parts of the loader are async, so the types will fit
together more nicely.