[DEV-8616] Merge ProgramQuoteCleaner into loader
Note that this is a near-verbatim copy, with minor modifications to get it
to compile with TS.  Further improvements will be made at some point.

There's no use in having all of these similar concepts in different
places.  The convention with OOP was to have each class in its own separate
file, but functional programs don't have quite the same distinction.

It'll be easier to reason about the system if all of this is the in the same
place.