[DEV-8616] Program: Add whenq and qstep properties
These are dynamically generated (and then cached), and will only be used by
the server.  I wanted to avoid generating them as part of Program.js (from
liza-proguic) because that'd bloat the JS sent to the client even more when
the information is not only redundant, but will also (at present) not be
used by the client.

The note about ease.js here represents a bug that must have existed since
getters were first introduced nearly a decade ago.  I don't have the time or
will to fix it, since we're actively deprecating ease.js.