[DEV-8616] Server: Persist quote state after migration
Migration may cause a kickback, so this will ensure that it is immediately
persisted.

This also logs information about whether a kickback has occurred (just a
little bit).  In the future I'd like to also log what fields caused it.

Tests are absent for this.  This has been long delayed and certain things
have to give; I have more important things I have to move on to.