[DEV-8616] loader: Kick back to newly applicable on program version migration
See extensive detail in the unit tests and comments.  This is the core of
this task and the feature that's been worked toward all along.

Also note that this forces the user to the current step when loading a quote
now, which will break hyperlinks to specific steps.  I don't think that's
currently done, and in practice we redirect users to policies after
documents (quotes) are locked, so this shouldn't affect us; ideally, though,
this would use e.g. the kickback action, which does not currently work
because the Quote object is not yet available to the action on the client
during init.

This task is also not all that great in terms of FP, but as was mentioned in
a previous commit, I intend to refactor further later on.  We also need time
to get the team acquainted before diving too deep into a sea of monads.