[DEV-8616] Client: Force validations on step load
The intent of this is to catch validations caused from server-side changes,
since we previously assumed (incorrectly) that, once things are in the
bucket, they are in a valid state, and forever will be.

This is not ideal, but is an incremental step toward a new way of
considering bucket data, which will require a far more fundamental
refactoring.