[DEV-8616] loader: Invalidate fields on kickback
This is a step toward implementing states for bucket field values (the
intent will be something along the lines of
NA|Pending v|Invalid v|Valid v).

Since I do not have time to fundamentally rework the Bucket and how it's
accessed at the moment, this does the best we can with what we have by
introducing the concept of an invalid state using an ASCII NACK character as
a string prefix.  The validators will consider this to be invalid in a
guaranteed way, and calculated values have been modified to take that into
consideration as well, since the validators may not have run by
then (e.g. on init).

eslint: Ignore loader, since it fails to parse

  0:0  error  Parsing error: Cannot read property 'map' of undefined

Will check again in future eslint versions.