[DEV-8616] Clear DOM on invalid state only on emptyBucket
This restores the DataValidator back to its previous behavior, because it
was causing unintended side-effects with displays.  Really, though, it
probably would have caused side-effects with fields as well by clearing them
in certain scenarios when a sibling index was modified.

This was a rushed change initially due to time constrains, as is this
change, so I want to be a bit more conservative and isolate the clearing as
much as possible.  Implementing this _properly_ would take far too long.

The comment in `emptyBucket` explains what this change does.  Essentially,
`undefined` should only be in the bucket on empty if the validator failed
for existing bucket data, so we should clear the field.  Clearing the field
ensures that default values on DOM elements are forcefully cleared so that
the state of the bucket (bogus) is accurately reflected, and the user is
able to select a value even if that value is the default.